Removed some methods of main class into a seperate QueryMYSQL class
